Move code from partAnalysis to partFuncs
